+++ system/src/Grav/Common/Cache.php
@@ -65,7 +65,8 @@ class Cache extends Getters
     protected static $all_remove = [
         'cache://',
         'cache://images',
-        'asset://'
+        'asset://',
+        'tmp://'
     ];
 
     protected static $assets_remove = [
@@ -78,6 +79,10 @@ class Cache extends Getters
 
     protected static $cache_remove = [
         'cache://'
+    ];
+
+    protected static $tmp_remove = [
+        'tmp://'
     ];
 
     /**
@@ -183,7 +188,7 @@ class Cache extends Getters
 
             case 'memcached':
                 $memcached = new \Memcached();
-                $memcached->connect($this->config->get('system.cache.memcached.server', 'localhost'),
+                $memcached->addServer($this->config->get('system.cache.memcached.server', 'localhost'),
                     $this->config->get('system.cache.memcached.port', 11211));
                 $driver = new DoctrineCache\MemcachedCache();
                 $driver->setMemcached($memcached);
@@ -308,6 +313,9 @@ class Cache extends Getters
                 break;
             case 'cache-only':
                 $remove_paths = self::$cache_remove;
+                break;
+            case 'tmp-only':
+                $remove_paths = self::$tmp_remove;
                 break;
             default:
                 $remove_paths = self::$standard_remove;