+++ system/src/Grav/Common/Config/Setup.php
@@ -17,6 +17,8 @@ use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
 
 class Setup extends Data
 {
+    public static $environment;
+
     protected $streams = [
         'system' => [
             'type' => 'ReadOnlyStream',
@@ -26,6 +28,7 @@ class Setup extends Data
         ],
         'user' => [
             'type' => 'ReadOnlyStream',
+            'force' => true,
             'prefixes' => [
                 '' => ['user'],
             ]
@@ -78,6 +81,7 @@ class Setup extends Data
         ],
         'cache' => [
             'type' => 'Stream',
+            'force' => true,
             'prefixes' => [
                 '' => ['cache'],
                 'images' => ['images']
@@ -85,14 +89,23 @@ class Setup extends Data
         ],
         'log' => [
             'type' => 'Stream',
+            'force' => true,
             'prefixes' => [
                 '' => ['logs']
             ]
         ],
         'backup' => [
             'type' => 'Stream',
+            'force' => true,
             'prefixes' => [
                 '' => ['backup']
+            ]
+        ],
+        'tmp' => [
+            'type' => 'Stream',
+            'force' => true,
+            'prefixes' => [
+                '' => ['tmp']
             ]
         ],
         'image' => [
@@ -120,7 +133,7 @@ class Setup extends Data
      */
     public function __construct($container)
     {
-        $environment = $container['uri']->environment() ?: 'localhost';
+        $environment = static::$environment ?: ($container['uri']->environment() ?: 'localhost');
 
         // Pre-load setup.php which contains our initial configuration.
         // Configuration may contain dynamic parts, which is why we need to always load it.
@@ -194,9 +207,13 @@ class Setup extends Data
             if (isset($config['paths'])) {
                 $locator->addPath($scheme, '', $config['paths']);
             }
+
+            $override = isset($config['override']) ? $config['override'] : false;
+            $force = isset($config['force']) ? $config['force'] : false;
+
             if (isset($config['prefixes'])) {
                 foreach ($config['prefixes'] as $prefix => $paths) {
-                    $locator->addPath($scheme, $prefix, $paths);
+                    $locator->addPath($scheme, $prefix, $paths, $override, $force);
                 }
             }
         }