+++ system/src/Grav/Common/GPM/Installer.php
@@ -88,8 +88,8 @@ class Installer
 
         $zip = new \ZipArchive();
         $archive = $zip->open($package);
-        $cache_dir = Grav::instance()['locator']->findResource('cache://', true);
-        $tmp = $cache_dir . DS . 'tmp/Grav-' . uniqid();
+        $tmp_dir = Grav::instance()['locator']->findResource('tmp://', true, true);
+        $tmp = $tmp_dir . '/Grav-' . uniqid();
 
         if ($archive !== true) {
             self::$error = self::ZIP_OPEN_ERROR;