+++ system/src/Grav/Common/Page/Medium/StaticImageMedium.php
@@ -23,6 +23,6 @@ class StaticImageMedium extends Medium
     {
         empty($attributes['src']) && $attributes['src'] = $this->url($reset);
 
-        return [ 'name' => 'image', 'attributes' => $attributes ];
+        return [ 'name' => 'img', 'attributes' => $attributes ];
     }
 }