+++ system/src/Grav/Console/Gpm/SelfupgradeCommand.php
@@ -49,6 +49,9 @@ class SelfupgradeCommand extends ConsoleCommand
      */
     private $upgrader;
 
+    protected $all_yes;
+    protected $overwrite;
+
     /**
      *
      */
@@ -69,6 +72,12 @@ class SelfupgradeCommand extends ConsoleCommand
                 InputOption::VALUE_NONE,
                 'Assumes yes (or best approach) instead of prompting'
             )
+            ->addOption(
+                'overwrite',
+                'o',
+                InputOption::VALUE_NONE,
+                'Option to overwrite packages if they already exist'
+            )
             ->setDescription("Detects and performs an update of Grav itself when available")
             ->setHelp('The <info>update</info> command updates Grav itself when a new version is available');
     }
@@ -79,6 +88,8 @@ class SelfupgradeCommand extends ConsoleCommand
     protected function serve()
     {
         $this->upgrader = new Upgrader($this->input->getOption('force'));
+        $this->all_yes = $this->input->getOption('all-yes');
+        $this->overwrite = $this->input->getOption('overwrite');
 
         $this->displayGPMRelease();
 
@@ -89,7 +100,6 @@ class SelfupgradeCommand extends ConsoleCommand
         $release = strftime('%c', strtotime($this->upgrader->getReleaseDate()));
 
         if (!$this->upgrader->meetsRequirements()) {
-            $this->output->writeln("");
             $this->output->writeln("<red>ATTENTION:</red>");
             $this->output->writeln("   Grav has increased the minimum PHP requirement.");
             $this->output->writeln("   You are currently running PHP <red>" . PHP_VERSION . "</red>, but PHP <green>" . GRAV_PHP_MIN . "</green> is required.");
@@ -100,8 +110,16 @@ class SelfupgradeCommand extends ConsoleCommand
             exit;
         }
 
-        if (!$this->upgrader->isUpgradable()) {
+        if (!$this->overwrite && !$this->upgrader->isUpgradable()) {
             $this->output->writeln("You are already running the latest version of Grav (v" . $local . ") released on " . $release);
+            exit;
+        }
+
+        Installer::isValidDestination(GRAV_ROOT . '/system');
+        if (Installer::IS_LINK === Installer::lastErrorCode()) {
+            $this->output->writeln("<red>ATTENTION:</red> Grav is symlinked, cannot upgrade, aborting...");
+            $this->output->writeln('');
+            $this->output->writeln("You are currently running a symbolically linked Grav v" . $local . ". Latest available is v". $remote . ".");
             exit;
         }
 
@@ -109,12 +127,12 @@ class SelfupgradeCommand extends ConsoleCommand
         new ArrayInput([]);
 
         $questionHelper = $this->getHelper('question');
-        $skipPrompt = $this->input->getOption('all-yes');
+
 
         $this->output->writeln("Grav v<cyan>$remote</cyan> is now available [release date: $release].");
         $this->output->writeln("You are currently using v<cyan>" . GRAV_VERSION . "</cyan>.");
 
-        if (!$skipPrompt) {
+        if (!$this->all_yes) {
             $question = new ConfirmationQuestion("Would you like to read the changelog before proceeding? [y|N] ",
                 false);
             $answer = $questionHelper->ask($this->input, $this->output, $question);
@@ -177,8 +195,8 @@ class SelfupgradeCommand extends ConsoleCommand
      */
     private function download($package)
     {
-        $cache_dir = Grav::instance()['locator']->findResource('cache://', true);
-        $this->tmp = $cache_dir . DS . 'tmp/Grav-' . uniqid();
+        $tmp_dir = Grav::instance()['locator']->findResource('tmp://', true, true);
+        $this->tmp = $tmp_dir . '/Grav-' . uniqid();
         $output = Response::get($package['download'], [], [$this, 'progress']);
 
         Folder::mkdir($this->tmp);